Making final changes on the project
