+++ package.json
@@ -1,51 +1,43 @@
 {
-  "name": "javascript-Capstone",
+  "name": "javascript-capstone",
   "version": "1.0.0",
-  "description": "",
-  "private": true,
+  "description": "<a name=\"readme-top\"></a>",
+  "main": "webpack.config.js",
+  "directories": {
+    "test": "tests"
+  },
   "scripts": {
-    "test": "jest --detectOpenHandles",
+    "test": "jest --coverage --env=jsdom",
     "build": "webpack",
-    "dev": "webpack serve"
+    "watch": "webpack --watch",
+    "start": "webpack serve --open"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.22.1",
-    "@babel/preset-env": "^7.22.4",
+    "@babel/plugin-transform-modules-commonjs": "^7.22.5",
+    "@babel/preset-env": "^7.22.9",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^29.5.0",
     "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
-    "eslint": "^8.45.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.28.0",
-    "html-webpack-plugin": "^5.5.3",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.27.5",
-    "hint": "^7.1.8",
-    "i": "^0.3.7",
-    "jest": "^29.5.0",
-    "jest-localstorage-mock": "^2.4.26",
-    "npm": "^9.6.7",
-    "sass": "^1.62.1",
-    "sass-loader": "^13.3.1",
+    "hint": "^7.1.10",
+    "html-webpack-plugin": "^5.5.3",
+    "sass": "^1.64.2",
+    "sass-loader": "^13.3.2",
     "style-loader": "^3.3.3",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0",
-    "webpack": "^5.85.0",
-    "webpack-cli": "^5.1.3",
-    "webpack-dev-server": "^4.15.0"
+    "webpack-dev-server": "^4.15.1"
   },
   "dependencies": {
     "axios": "^1.4.0",
-    "html-webpack-plugin": "^5.5.1",
-    "jsdom": "^22.1.0",
-    "jest": "^29.6.2",
-    "lodash": "^4.17.21"
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4"
   }
 }
