+++ src/index.html
@@ -1,46 +1,50 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css"
+      integrity="sha512-+4zCK9k+qNFUR5X+cKL9EIR+ZOhtIloNl9GIKS57V1MyNsYpYcUrUeQc9vNfzsWfV28IaLL3i96P9sdNyeRssA=="
+      crossorigin="anonymous"
+    />
+    <title>Suleiman Movies</title>
+  </head>
 
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css"
-    integrity="sha512-+4zCK9k+qNFUR5X+cKL9EIR+ZOhtIloNl9GIKS57V1MyNsYpYcUrUeQc9vNfzsWfV28IaLL3i96P9sdNyeRssA=="
-    crossorigin="anonymous" />
-  <title>
-    <%= htmlWebpackPlugin.options.title %>
-  </title>
-</head>
-
-<body>
-  <header class="head">
-    <section class="container">
-      <nav class="navbar">
-        <h2 class="logo">Vidvib</h2>
-        <ul class="navbar-ul">
-          <li class="navbar-list">
-            <a href="" class="navbar-link">Home</a>
-          </li>
-          <li class="navbar-list">
-            <a href="" class="navbar-link">Movies</a>
-          </li>
-          <li class="navbar-list">
-            <a href="" class="navbar-link">Series</a>
-          </li>
-        </ul>
+  <body>
+    <header>
+      <nav>
+        <div class="nav-menu">
+          <h2>Suleiman Movies</h2>
+          <ul>
+          <ul class="nav-list">
+            <li>
+              <a href="#" class="mobile-links">Movies</a>
+            </li>
+            <li>
+              <a href="#" class="mobile-links">Series</a>
+            </li>
+            <li>
+              <a href="#" class="mobile-links">Contact</a>
+            </li>
+          </ul>
+        </div>
       </nav>
-    </section>
-  </header>
-  <main class="movies">
-    <section class="container display-movies"></section>
-  </main>
-  <footer class="foot">
-    <section class="container">
-    <h2>vidvib</h2>
-    <p>&copy; Created by Newton and Suleiman @Microverse under cc licence</p>
-  </section>
-  </footer>
-</body>
-
+    </header>
+    <div class="container">
+      <p class="msg"></p>
+    </div>
+    <main class="movies">
+      <section class="container display-movies"></section>
+      <section class="popContainer"></section>
+    </main>
+    <footer>
+      <section class="container">
+        <h2>Suleiman</h2>
+        <p>&copy; Designed by Newton and Suleiman</p>
+      </section>
+    </footer>
+  </body>
 </html>
