+++ src/index.js
@@ -1,5 +1,4 @@
-/* eslint-disable import/no-named-as-default-member */
 import displayList from './modules/display-movies';
-import './styles.css';
+import './styles/style.css';
 
 displayList();
