+++ src/modules/commentCount.js
@@ -1,9 +1,9 @@
 const commentCounter = async () => {
-    const commentsContainer = document.querySelector('.commentDiv');
-    const commentCount = commentsContainer.children.length;
-    const counterElement = document.querySelector('.counter');
-    if (counterElement) {
-      counterElement.textContent = `(${commentCount})`;
-    }
-  };
-  export default commentCounter;
+  const commentsContainer = document.querySelector('.commentDiv');
+  const commentCount = commentsContainer.children.length;
+  const counterElement = document.querySelector('.counter');
+  if (counterElement) {
+    counterElement.textContent = `(${commentCount})`;
+  }
+};
+export default commentCounter;
