--- src/modules/localStorage.js
@@ -1,6 +1,0 @@
-const getAppName = () => {
-  const appName = '47awMaoQqtIaDLtHDdVW';
-  return appName;
-};
-
-export default getAppName;
