--- src/styles.css
@@ -1,108 +1,0 @@
-html {
-  scroll-behavior: smooth;
-}
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-i {
-  cursor: pointer;
-}
-
-a {
-  text-decoration: none;
-}
-
-img {
-  width: 100%;
-}
-
-.like-icon {
-  width: 20px;
-  cursor: pointer;
-}
-
-button {
-  padding: 10px;
-  border-radius: 5px;
-  font-size: 16px;
-  color: #fff;
-  background: blue;
-  border: none;
-}
-
-.container {
-  padding: 0 40px;
-  max-width: 1440px;
-  margin: 0 auto;
-}
-
-/* header */
-.head {
-  background: yellow;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 50px;
-}
-
-.navbar .navbar-ul {
-  list-style-type: none;
-  display: flex;
-  gap: 20px;
-}
-
-.navbar .navbar-ul .navbar-link {
-  color: #fff;
-  font-size: 15px;
-  font-weight: 600px;
-}
-
-.movies {
-  margin: 50px 0;
-}
-
-.movies .container {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-}
-
-.movies .card {
-  max-width: 400px;
-  margin: auto;
-}
-
-.movies .card h3 {
-  margin-block: 10px;
-}
-
-.movies .card .card-body {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-block: 20px;
-}
-
-/* Footer  */
-.foot {
-  background-color: grey;
-}
-
-.foot .container {
-  height: 50px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  color: white;
-}
