+++ src/styles/style.css
@@ -1,0 +1,335 @@
+* {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  scroll-behavior: smooth;
+  box-sizing: border-box;
+}
+
+li {
+  list-style: none;
+}
+
+a {
+  text-decoration: none;
+}
+
+nav {
+  position: relative;
+}
+
+@media (min-width: 375px) {
+  .nav-menu {
+    display: block;
+  }
+
+  .mobile-menu-icon {
+    display: flex;
+    justify-content: end;
+    padding: 20px 35px;
+  }
+
+  .mobile-menu-close-icon {
+    display: none;
+  }
+
+  .social-icons ul {
+    display: flex;
+    gap: 20px;
+    margin-top: 1.25rem;
+    list-style: none;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .social-icons ul li a {
+    text-decoration: none;
+  }
+
+  .nav-menu.show {
+    display: block;
+    height: 100vh;
+    background: #fff;
+    animation: 2s slideIn ease-in alternate;
+  }
+
+  @keyframes slideIn {
+    0% {
+      transform: translateX(-50%);
+    }
+
+    100% {
+      transform: translateX(0);
+    }
+  }
+
+  .nav-menu.show ul {
+    margin-left: 28px;
+    list-style: none;
+    display: grid;
+    gap: 40px;
+    padding-top: 108px;
+  }
+
+  .nav-menu.show li a {
+    font-weight: 600;
+    font-size: 32px;
+    line-height: 44px;
+    display: flex;
+    align-items: center;
+    color: #3a4a42;
+  }
+
+  .mobile-menu-icon.show {
+    display: none;
+  }
+
+  .mobile-menu-close-icon.show {
+    display: block;
+    position: absolute;
+    content: '';
+    top: 1.125rem;
+    right: 2.1875rem;
+  }
+
+  nav svg {
+    height: 60px;
+  }
+}
+
+@media (min-width: 768px) {
+  .nav-menu {
+    display: flex;
+    position: relative;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem 5rem;
+  }
+
+  .nav-menu ul {
+    display: flex;
+    gap: 25px;
+  }
+
+  .nav-menu ul li a {
+    text-decoration: none;
+    font-family: 'Inter', sans-serif;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 20px;
+    line-height: 32px;
+    text-align: center;
+    color: #050e1d;
+  }
+
+  .nav-menu li a:hover {
+    color: #2b45bb;
+  }
+
+  .nav-menu li a:active {
+    color: #5c6cb3;
+  }
+
+  .mobile-menu-icon {
+    display: none;
+  }
+
+  .mobile-menu-close-icon {
+    display: none;
+  }
+}
+
+img {
+  width: 100%;
+}
+
+.like-icon {
+  width: 1.25rem;
+  cursor: pointer;
+}
+
+button {
+  padding: 10px 15px;
+  font-size: 1rem;
+  color: #fff;
+  cursor: pointer;
+  background-color: green;
+  border: none;
+  margin-left: 0.25rem;
+}
+
+button:hover {
+  color: green;
+  background-color: white;
+}
+
+.movie h3 {
+  margin-left: 0.25rem;
+}
+
+.msg {
+  display: none;
+  position: fixed;
+  left: 70%;
+  width: 200px;
+  padding: 10px 10px;
+  border-radius: 15px;
+  text-align: center;
+  color: #fff;
+  z-index: 99;
+}
+
+.container {
+  padding: 0 40px;
+  max-width: 1440px;
+  margin: 0 auto;
+}
+
+.movies {
+  margin: 3.125rem 0;
+}
+
+footer .container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 50px;
+}
+
+.movies .container {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 20px;
+}
+
+.movies .card {
+  background: #e5e5e5;
+  width: 20rem;
+  padding: 0.625px;
+  border-radius: 10px;
+}
+
+.movies img {
+  border-radius: 10px 10px 0 0;
+  overflow: hidden;
+}
+
+.movies .card .card-body {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-block: 20px;
+}
+
+.popContainer {
+  position: fixed;
+  display: none;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: 1;
+  padding: 60px 0;
+  overflow: auto;
+  background: rgba(36, 36, 36, 0.9);
+}
+
+.pop {
+  max-width: 600px;
+  padding: 20px;
+  margin: 0 auto;
+  background: #fff;
+  border-radius: 20px;
+}
+
+.popup {
+  display: flex;
+  flex-direction: column;
+}
+
+.comment-count {
+  display: flex;
+  gap: 20px;
+  margin-top: 10px;
+  margin-bottom: -10px;
+}
+
+.commentDiv {
+  height: 200px;
+  overflow-y: auto;
+  margin-block: 20px;
+}
+
+.commentDiv p {
+  background: rgba(36, 36, 36, 0.9);
+  color: #fff;
+  padding: 10px 20px;
+  margin-block: 10px;
+  border-radius: 20px;
+}
+
+.nameField,
+.commentField {
+  width: 100%;
+  border-radius: 20px;
+  padding: 5px 15px;
+  color: #242424;
+  font-size: 14px;
+  border: 2px solid #242424e6;
+}
+
+.nameField::placeholder,
+.commentField::placeholder {
+  color: #242424;
+}
+
+.closeBtn {
+  width: 20px;
+  margin-bottom: 10px;
+  position: relative;
+  left: 97%;
+}
+
+.addComment {
+  display: inline-block;
+  width: 200px;
+  height: 40px;
+  color: #fff;
+  margin-top: 10px;
+  background: rgba(36, 36, 36, 0.9);
+  border-radius: 20px;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+}
+
+form input {
+  width: 200px;
+  height: 30px;
+  padding: 10px;
+  margin-block: 10px;
+}
+
+footer {
+  border-top: 2px solid;
+}
+
+@media (max-width: 766px) {
+  .movies .container {
+    grid-template-columns: repeat(1, 1fr);
+  }
+
+  .movies .card {
+    margin: auto;
+  }
+
+  .foot .container {
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+    text-align: center;
+    height: 100px;
+  }
+}
