+++ tests/__mocks__/axios.js
@@ -1,0 +1,3 @@
+export default {
+  get: jest.fn(() => Promise.resolve({ data: null })),
+};
