--- tests/_mocks_/axios.js
@@ -1,3 +1,0 @@
-export default {
-    get: jest.fn(() => Promise.resolve({ data: null })),
-  };
