+++ tests/commentCounter.test.js
